Fix: mlx m1 호환되게끔 한줄 수정

+++ minilibx_mms_20210621/interface.swift
@@ -270,7 +270,7 @@ public func mlx_mouse_move_swift(_ winptr:UnsafeRawPointer, _ x:Int32, _ y:Int32
 	var pt = CGPoint(x:CGFloat(x), y:CGFloat(y))
 	pt = win.convertToDisplay(pt)
 	CGWarpMouseCursorPosition(pt)
-	CGAssociateMouseAndMouseCursorPosition(UInt32(1))
+	CGAssociateMouseAndMouseCursorPosition(1)
 	return Int32(0);
 }
